Linkedin SC/requirements.txt
playwright>=1.40.0
pandas>=2.0.0
python-dotenv>=1.0.0
asyncio-throttle>=1.0.0
openai>=1.0.0
aiohttp>=3.8.0
aiofiles>=23.0.0
Pillow>=10.0.0
openpyxl>=3.1.0
requests>=2.31.0
fastapi>=0.111.0
uvicorn>=0.30.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
